+++ CHANGES.md
@@ -1,5 +1,14 @@
 Change Log
 ==========
+
+### 1.39 - 2017-11-??
+
+* Added two new properties to `ImageryLayer` that allow for adjusting the texture sampler used for up- and down-sampling
+  of image tiles, namely `minificationFilter` and `magnificationFilter` with possible values `LINEAR`
+  (the default) and `NEAREST` defined in `TextureMinificationFilter` and `TextureMagnificationFilter`.
+  [#5846](https://github.com/AnalyticalGraphicsInc/cesium/issues/5846)
+    
+
 ### 1.38 - 2017-10-02
 
 * Breaking changes